Make image_getter more robust and faster.
Now it can download multiple large batches without crashing.

+++ common/imagenet_server/downloader.py
@@ -6,6 +6,7 @@ import collections
 import logging
 import multiprocessing
 import random
+import signal
 import time
 
 import data_augmentation
@@ -37,7 +38,7 @@ class DownloaderProcess(multiprocessing.Process):
     # Download the image.
     image = images.download_image(self.__url, keep_color=True)
     if image is None:
-      logging.warning("Failed to download %s." % (self.__url))
+      logger.warning("Failed to download %s." % (self.__url))
       self.__image_queue.put((self.__synset, self.__number, self.__url, None,
                               None))
       return
@@ -46,7 +47,11 @@ class DownloaderProcess(multiprocessing.Process):
     patches = data_augmentation.extract_patches(image)
 
     # Save the image to the queue.
-    logging.debug("Saving image: %s_%s" % (self.__synset, self.__number))
+    logger.debug("Saving image: %s_%s" % (self.__synset, self.__number))
+    # Defer acting on a terminate signal until after the queue call. (Which can
+    # block, but won't indefinitely as long as update() is being called
+    # regularly.)
+    signal.signal(signal.SIGTERM, signal.SIG_IGN)
     self.__image_queue.put((self.__synset, self.__number, self.__url, image,
                             patches))
 
@@ -68,7 +73,7 @@ class DownloadManager(object):
   """ Deals with managing and dispatching downloads. """
 
   def __init__(self, process_limit, disk_cache, mem_buffer,
-               patch_separation=-1, timeout=10):
+               patch_separation=-1, timeout=15):
     """
     Args:
       process_limit: Maximum number of downloads we can run at one time.
@@ -88,8 +93,10 @@ class DownloadManager(object):
     self.__download_queue = collections.deque()
     # Queue other processes use to communicate image data.
     self.__image_queue = multiprocessing.Queue()
-    # List of failed downloads.
-    self.__failures = []
+    # Set of failed downloads.
+    self.__failures = set([])
+    # Set of downloads that were terminated.
+    self.__terminated = set([])
 
   def download_new(self, synset, number, url):
     """ Registers a new image to be downloaded.
@@ -111,9 +118,11 @@ class DownloadManager(object):
     for process in multiprocessing.active_children():
       if (time.time() - process.get_start_time()) > self.__timeout:
         logger.warning("Terminating downloader process that took too long.")
-        process.terminate()
-        # Add it to the failures list, because it counts.
-        self.__failures.append(process.get_info())
+        if process.get_info() not in self.__terminated:
+          process.terminate()
+          # Add it to the failures list, because it counts.
+          self.__failures.add(process.get_info())
+          self.__terminated.add(process.get_info())
 
     processes = True
     while len(multiprocessing.active_children()) < self.__process_limit:
@@ -133,9 +142,14 @@ class DownloadManager(object):
       data = False
     while not self.__image_queue.empty():
       synset, name, url, image, patches = self.__image_queue.get()
+
+      if (synset, name, url) in self.__terminated:
+        # If it was terminated, this was already reported as a failure.
+        continue
+
       if image is None:
         # Download failed.
-        self.__failures.append((synset, name, url))
+        self.__failures.add((synset, name, url))
         continue
 
       if self.__patch_separation < 0:
@@ -149,6 +163,9 @@ class DownloadManager(object):
       # Add it to the disk cache.
       self.__disk_cache.add(image, name, synset)
 
+    if not (processes or data):
+      # We're done downloading everything, so we can clear the terminated set.
+      self.__terminated = set([])
     return processes or data
 
   def wait_for_downloads(self):
@@ -159,11 +176,11 @@ class DownloadManager(object):
     while self.update():
       time.sleep(1)
 
-    return self.__failures
+    return self.get_failures()
 
   def get_failures(self):
     """ Returns: The list of failures up until now. Every call to this function
     will clear the list after returning it. """
-    failures = self.__failures[:]
-    self.__failures = []
+    failures = list(self.__failures)
+    self.__failures = set([])
     return failures
